Provisioning playbooks should just setup provisioning
The docs mention you can use these on an existing box, which was wrong:
they include the katello and foreman playbooks.  So I had a box running
Katello in a specific way I wanted, and then ran the Katello
provisinoing playbook and it screwed it all up since it tried to install
Katello with the defaults, instead.

Provisoning playbooks should just do provisioning setup on an existing
box.